+++ CHANGES.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-07-18
+
 ### Added
 - New option: "Include Characters", to include/exclude certain characters from the world
   if players don't want to play as them.
@@ -216,7 +218,8 @@ objects for better organization and readability.
     - Hooks into Brotato to add the received items listed above and detect when
       locations are checked.
 
-[unreleased]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.0.6...v0.1.0
 [0.0.6]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/SpenserHaddad/Brotato-ArchipelagoClient/compare/v0.0.4...v0.0.5